rebase to the latest dialog
