+++ apps/designer/app/dashboard/projects/index.ts
@@ -1,0 +1,1 @@
+export { Projects } from "./projects";
