+++ apps/designer/app/dashboard/projects/projects.tsx
@@ -1,7 +1,7 @@
 import { Flex } from "@webstudio-is/design-system";
 import { EmptyState } from "./empty-state";
 import { Panel } from "./panel";
-import { Heading } from "./heading";
+import { Heading } from "../heading";
 import type { DashboardProject } from "@webstudio-is/prisma-client";
 import { ProjectCard } from "./project-card";
 import { CreateProject } from "./create-project";
